% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{faf4_interregional_distances}
\alias{faf4_interregional_distances}
\title{FAF inter-regional distances}
\format{A data frame with 17414 rows and 3 variables:
\describe{
  \item{dms_orig}{Origin domestic FAF region}
  \item{dms_dest}{Destination domestic FAF region}
  \item{distance}{Roadway distance between origin and destination, in miles}
}}
\source{
Gregory Macfarlane
}
\usage{
faf4_interregional_distances
}
\description{
A dataset containing the distances between the centroids of FAF regions,
skimmed from the National Highway Planning Network for FAF version 4 regions
}
\keyword{datasets}

